Use the same ssh key as ubuntu-device-flash does
This patch makes snapcraft locate the same ssh key that a corresponding
bit of code in ubuntu-device-flash is looking for, so that the
subsequent ssh and scp commands can use that key (identity) explicitly.

This fixes an issue where the user has multiple keys and due the
mismatch between the two programs, snapcraft run cannot copy or install
the package inside the started virtual machine.

Fixes: https://bugs.launchpad.net/snapcraft/+bug/1486659
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>