Don't try to login interactively to snapcraft run vms
This patch fixes a potential error case where snapcraft run misbehaves
and each of the invoked ssh or scp commands is blocked by interactive
ssh authentication (password prompt) that is typically totally hidden in
the large amount of output produced by the booting machine.

To prevent this from happening the KbdInteractiveAuthentication ssh
option is set to "no".

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>