Add support for SNAPCRAFT_RUN_QEMU_ARGS
This patch extends snapcraft run to respect the environment variable
SNAPCRAFT_RUN_QEMU_ARGS. This optional environment variable, when
present, is parsed as shell would parse it and appended to the
invocation of qemu/kvm.

This allows the developer to provide additional arguments to qemu. This
can be used, for example, to pass through USB devices from the host.
This can enable a lot of hardware-specific use cases directly inside the
snapcraft run workflow.

For example:

$ export SNAPCRAFT_RUN_QEMU_ARGS="-usb -device usb-host,hostbus=1,hostaddr=10"
$ snapcraft run

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>