Add 'simple' and 'has-leftovers' flags to all job units
This patch adds two flags to each job (a plainbox concept) used for
snapcraft integration tests.

The 'simple' flag indicates that some more elaborate job requirements
are relieved so that a minimal job definition can be shorter. In
practical terms description and estimated_duration are no longer
recommended when they are absent. This fits in nicely with the concept
of "unit tests" that are typically short enough not to warrant a
docstring or other verbose documentation.

The 'has-leftovers' flag toggles the default behavior of warning that
the command executed for a given job has left some files over in the
temporary directory in which it was executed. The goal of leftover files
was to catch mistakes in an environment where jobs were rarely creating
files, and if any, they were expected to put them in carefully-selected
directories. Snapcraft is a heavy "polluter" of the temporary directory
so this flag just makes test execution less noisy so that actual
problems are easier to spot.

This feature is harmless on earlier versions of plainbox (flags are
silently ignored). With the development release it works as expected.

Fixes: https://bugs.launchpad.net/snapcraft/+bug/1484596
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>