Get deps from the part's sourcedir, fail if needed
When grabbing deps, there is still nothing in the part's builddir
so we need to search for them in sourcedir for the part.

Missing a package.xml logged a warning and marked the pull succesful,
this brought dire consequences when building as nothing required for
a build existed and a pull --force was always required (silent error).
This seems counterproductive, so failing the pull step seems reasonable.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>