Clean for parts and remove dup from lf and plugin.
Parts can now be individually cleaned up, involved moving some
of the logic to the plugin handler in the lifecycle itself.

Define things in one place only and expose them accordingly.

sourcedir, builddir and installdir are handled in the plugin while
stagedir, snapdir, ubuntudir and the state file by the lifecycle.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>