Enabling testing with new cli
This inlcudes fixes to how recursion is handled, removes obsolete tests and
does some minor clean ups on some of the code paths that required change to
make them work. The minimum python version has been bumped to 3.5 to match
what is to be released with xenial.

Given the cyclic dep introduced, the PluginHandler and its supporting
functions have been moved to a new snapcraft.pluginhandler module.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>