Change in test_wiki.py
There is a line(#50) that is not tested by the test-code. In the added test we try to be sure that 'type' part of the part was deleted successfully. To do so, we check that the result of get_part function is not equal to the result where source-type part is present. In the case of part-not-in-wiki, we should check also that source-type isn't there so result must be 'None'