Catkin plugin: Improve cmake path rewrite.
Currently the Catkin plugin prepares to build by rewriting the
Catkin <package>Config.cmake files from the ROS Ubuntu packages
to point to the install directory instead of system paths. However,
it was hard-coded to only rewrite a specific path, and misses others.
This commit makes the path rewrites more generic and robust.

Signed-off-by: Kyle Fazzari <kyle@canonical.com>