Fix environment variable handling
Environment variables now correctly point to stagedir and snapdir and
don't expose the internal installdir paths for a part.

This required an intermediate change in the handling of stage-packages
but it also gives it a semantic meaning by unpacking everything into
the staging directory. Future work can make the handling of
stage-package more transactional or in a single shot.

LP: #1531481