Copy entire source, even if source-subdir is specified.
Currently, if source-subdir is specified, Snapcraft copies
ONLY that subdir (and excludes the rest of the source) and
attempts to build from it. This only works for some projects,
not all. A better solution is to copy the entire source and
simply build from WITHIN the subdir. That is the change
introduced here.

LP: #1549676

Signed-off-by: Kyle Fazzari <kyle@canonical.com>