Support parallel builds (enabled by default).
Several plugins within Snapcraft support parallel builds, but
Snapcraft currently doesn't take advantage of this and builds
them with a single job. This commit adds support for parallel
builds, and enables them by default in the autotools, cmake,
and make plugins. It also adds an option to disable parallel
builds via the `--no-parallel-build` option for the case of
running in a RAM-limited environment.

LP: #1551417

Signed-off-by: Kyle Fazzari <kyle@canonical.com>