travis: run the tests in the GCE trusty machine
We applied a hack in travis to get a xenial LXC and run the tests in
there. Something changed and now the tests are taking an excessive
amount of time, and timing out. We no longer need this tests to run in
xenial because we are already running them in Jenkins, so we can get rid
of the hack.

LP: #1555386