Support cleaning build step.
This commit adds the ability to clean only a part's built files.
This involves adding a new clean_build() method to the plugin API
as there are a few plugins that have more involved build steps
than the base implementation's.

This commit also reimplements clean_build() for the Go plugin.

LP: #1537786

Signed-off-by: Kyle Fazzari <kyle@canonical.com>