Support cleaning pull step.
This commit adds the ability to clean only a part's pulled files.
This involves adding a new clean_pull() method to the plugin API
as there are a few plugins that have more involved pull steps
than the base implementation's.

This commit also reimplements clean_pull() for the Go, NodeJS,
and Catkin plugins.

LP: #1537786

Signed-off-by: Kyle Fazzari <kyle@canonical.com>