Copy plugin: Add support for sources.
Currently the copy plugin can only be used for local assets and
pulling down build- or stage-packages. With this commit, the copy
plugin can also be used for remote sources with no build system.

This also means that the functionality contained within the
tar-content plugin is now duplicated in the copy plugin, so this
commit deprecated the tar-content plugin.

LP: #1559154

Signed-off-by: Kyle Fazzari <kyle@canonical.com>