Suggest a full clean if step clean is missing state.
Currently if attempting to run a per-step clean on a tree made

dirty by a version of Snapcraft that didn't support per-step

cleaning, Snapcraft suggests trying to run that step again. This

isn't possible since the step has already been run, so it actually

requires a full clean. This commit updates the suggestions to

reflect this fact.



LP: #1561328



Signed-off-by: Kyle Fazzari <kyle@canonical.com>