Copy plugin: Follow symlinks if outside of snap.
Currently the copy plugin copies all symlinks (using a hard link to them if possible). It should do its best to ensure symlinks will be valid when snapped, however, and follow the ones that aren't. This commit does a relatively conservative check for invalid symlinks. The symlink will be followed if any of the following conditions are true:

  - The symlink is absolute.
  - The symlink is pointing outside of the part's install
    directory.

It's possible to still get broken symlinks into a snap using this plugin with certain special cases, but this should catch the majority of use-cases.

LP: #1532515

Signed-off-by: Kyle Fazzari <kyle@canonical.com>