Lifecycle: Check to see if prerequisite is built.
Currently, if one specifies a part to be built, the lifecycle
checks to see if that part has any dependencies. If so, it
immediately errors out if those dependencies are not in the list
to be built. This commit checks to make sure the dependencies are
not already built before erroring.

LP: #1537790

Signed-off-by: Kyle Fazzari <kyle@canonical.com>