Automatically re-strip if dirty.
Currently Snapcraft doesn't notice YAML changes if a given step has
already run-- it simply says that it already ran and halts execution.
This commit changes that for the strip step, where now if one changes
the `snap` keyword in the YAML, strip will notice that it's out of date
and automatically clean itself and re-run.

LP: #1477904

Signed-off-by: Kyle Fazzari <kyle@canonical.com>