Remove global state attributes from snapcraft.common (#434)
This includes work to move architecture and parallel build related
options to `snapcraft.ProjectOptions` in an effort to move global state
away. This work also includes messages to support users of the API
affected by the change and a transparent migration to include these
`project` options into the plugin attributes.

LP: #1551849

Signed-off-by: Sergio Schvezov <sergio.schvezov@ubuntu.com>
Signed-off-by: Kyle Fazzari <kyle@canonical.com>