Detect dirty build step.
Currently Snapcraft notices YAML changes only for the strip and stage
steps. For the other steps, if they've already run, it simply says so
and bails. This commit changes that for the build step, where now if one
changes any part of the YAML that the part's plugin says is associated
with building, build will notice that it's out of date and error out
asking the user to clean the build step to continue.

The build step is not automatically cleaned like the strip and stage
steps because builds can take a very long time and we don't want to blow
one away without the user explicitly asking to do so.

LP: #1477904

Signed-off-by: Kyle Fazzari <kyle@canonical.com>