Automatically re-stage if dirty. (#443)
Currently Snapcraft only notices YAML changes for the strip step. For
all other steps, if the step has already run it simply says so and
bails. This commit changes that for the stage step, where now if one
changes the `stage` keyword in the YAML, stage will notice that it's out
of date and automatically clean itself (along with the strip step since
it depends upon stage) and re-run.

LP: #1477904

Signed-off-by: Kyle Fazzari <kyle@canonical.com>