Detect dirty pull step.
Currently Snapcraft notices YAML changes for every step except the pull
step. This commit adds this functionality for the pull step, where now
if one changes any part of the YAML that the part's plugin says is
associated with pulling, pull will notice that it's out of date and
error out asking the user to clean the pull step to continue.

The pull step is not automatically cleaned like the strip and stage
steps because pulls can take a very long time (and require network) and
we don't want to blow one away without the user explicitly asking to do
so.

LP: #1477904

Signed-off-by: Kyle Fazzari <kyle@canonical.com>