Add support for confinement property.
Snaps need to be able to specify if they require devmode or if they can
be run confined. This will allow for snapd to provide reasonable errors
if one tries to install a snap that cannot run successfully under
confinement. This commit adds the "confinement" property, which is
required and must be one of "enabled" or "devmode." `snapcraft init`
generates a YAML that sets "confinement" to "devmode."

LP: #1580819

Signed-off-by: Kyle Fazzari <kyle@canonical.com>