Add support for confinement property. (#501)
Snaps need to be able to specify if they require devmode or if they can
be run confined. This will allow for snapd to provide reasonable errors
if one tries to install a snap that cannot run successfully under
confinement.

This commit adds the "confinement" property, which must be one of
"strict" or "devmode." This property is optional, and defaults to
"strict" if unspecified (to maintain backward-compatibility with current
behavior). However, `snapcraft init` generates a YAML that sets
"confinement" to "devmode" since we expect most snaps to start out
needing that.

LP: #1580819

Signed-off-by: Kyle Fazzari <kyle@canonical.com>