Avoid creating snapcraft-tour/snapcraft-… in loop
When launching multiple time the command, as we could have a destination
path where we either dump the code or create a snapcraft-tour subdirectory
if it exists, multiple snapcraft tour command would create a snapcraft-tour
folded hierarchy.

To avoid that, if we didn't specify any parameter, only try to create a
snapcraft-tour subdir and don't nest under it. Added a test for that.