Added mavent-targets for when building composite maven projects in snapcraft. Composite maven projects do not build to 'target'.  This update allows the user to specify the child target to copy.
parts:
  algolink-mvn:
    plugin: maven
    source: .
    maven-targets:
      - algolink-gui
      - algolink-cli

By default, the basedir/target directory is searched, otherwise it uses the 'target' directory of each specified value.
The installdir for each child project is the name of the child project.  If no values are specified, it reverts to the default of 'jar' or 'war' depending on the target type of the project.