Fix the store integration tests
The tests were failing because we added a new one that attempts to register
two snaps too fast, in order to check the error the store sends for too many
attempts. After the second registration, we need to wait a little so the
following tests can register without getting the same error.

LP: #1599268