Catkin plugin: Enforce C.UTF-8 locale. (#497)
ROS apps run fine on Ubuntu Core, but throw boost locale errors on
Xenial desktop. This is due to LP: #157641 where the desktop locales
aren't available to snaps even though the locale environment is set for
the snaps. This PR is a quick and dirty fix for LP: #1577557 that simply
sets `LC_ALL=C.UTF-8` for snaps with Catkin parts. This PR can be
reverted once LP: #1576411 is actually fixed.

LP: #1577557

Signed-off-by: Kyle Fazzari <kyle@canonical.com>