Implement basic form of `snapcraft register-key`
This adds a basic register-key command that can be used to register keys
from your GPG keyring.  At present, they must already be associated with
your Launchpad account, although after the work to move this to SSO is
completed we'll be able to add such associations inline in snapcraft as
well.

We don't yet have all the prescribed error modes in place, and nor are
we able to tell which keys are already enabled.  That requires some more
work on the store side.

LP: #1612730