Dump plugin: Don't remove install directory.
This prevents the usage of stage-packages. It was done because
shutil.copytree simply can't deal with merging two directories, but the
dump plugin needs to be able to deal with that situation.

This commit takes a lightly modified version of the  _linktree() logic
from the Copy plugin and puts it into the Dump plugin. This has the
added benefit of not copying broken symlinks into the snap as well.

LP: #1611827

Signed-off-by: Kyle Fazzari <kyle@canonical.com>