Adjust login to conform to UX specification (#804)
* Adjust login to conform to UX specification

The store now returns an unambiguous error to indicate that two-factor
authentication is required, so we can use that to make the login process
behave how the UX specification says it should: only prompt for 2fa if
necessary, and provide a hint on a successful login without 2fa.

One obvious hack in this change is that we don't yet provide a
convenient per-store link for enabling 2fa, but instead just hardcode a
link to the SSO FAQ for now.  At the moment, people have to explicitly
opt into 2fa by joining a Launchpad team, which is a deliberate decision
because we don't have a good recovery facility and can't afford the
support bandwidth of doing manual recovery for millions of users.  When
this changes then we can also update snapcraft to do something more
graceful here, but for the moment I think this is tolerable.

LP: #1621710

* Remove unused test path

* Test another /tokens/discharge response case