Add `snapcraft create-key`
Wrapping `snap create-key` means that we don't confuse users by
expecting them to use two similar-but-different commands during the key
creation and registration workflow.  In addition, this augments `snap
create-key` by failing if a key with the same name is already registered
with the store, since a later registration attempt would fail in that
case.