Add support for hooks.
Hooks are now supported in snapd and the click reviewer tools. Time to
support them in snapcraft as well. This is done declaratively via a
`hooks` option in the snapcraft.yaml, for example:

    hooks:
      configure:
        command: run-my-hook arg1 arg2
        plugs: [network]

This works exactly like apps, where `run-my-hook` should be a binary
available within the snap. Then during the prime step, snapcraft will
generate the hook (the configure hook in this example), which calls the
command supplied.

LP: #1586465

Signed-off-by: Kyle Fazzari <kyle@canonical.com>