Support download and validate on branded stores.
Allow users to set `SNAPCRAFT_UBUNTU_STORE` and perform operations
on branded stores.

So far it only affects interactions with the Snap Index (CPI), which
already support the 'X-Ubuntu-Store' header and further permission
checks.

It will be probably used as the foundation for supporting push and
release for branded stores, instead of discrete '--store'.

LP: #1650210