pluginhandler: add support for disabling system library migration (#989)
Currently snapcraft, during the prime step for a part, checks to see if
that part needs any libraries from the system. This is not always
desired, as in some cases the libraries will be provided by another snap
via the content interface.

This commit introduces a new property to the base part YAML:
build-attributes. It's simply a list of build attributes, only one of
which exists right now: `no-system-libraries`. For example:

    parts:
      # ...
      example:
        plugin: cmake
        build-attributes: [no-system-libraries]
      # ...

Using the `no-system-libraries` attribute has the effect of not pulling
any libraries from the system during the prime step for that part.

LP: #1639831

Signed-off-by: Kyle Fazzari <kyle@canonical.com>