gradle plugin: update gradle plugin to support both gradle and gradlew. (#1024)
Within snapcraft, the invocation of gradle v. ./gradlew is determined by the existence of the wrapper script. Same goes for if there is a build dependency which snapcraft must install.

The arguments passed to wrapper and to gradle are the same.

On top of these changes, I have added in the ability to specify what the target output directory is from gradle. Gradle has default output directories, but it is very flexible and is possible to change the output directory to be anything. Thus, snapping the data from the gradle plugin is not always easily done. The default is build/libs which works if nothing is altered.

LP: #1653182