Download all Python packages in the python plugin using one command.
Before, there was one command for downloading the packages in
requirements file, one for the ones in setup.py, and one for downloading
the manually specified packages.

That broke the case where you specify high-level dependencies in
setup.py without version-specific contstraints and use requirements.txt
to pin the packages to specific versions. The end result was that even
though packages were pinned in requirements.txt, snapcraft would
download the latest versions of packages.