schema,copy plugin: better errors when item has no value.
Snapcraft supports a few different objects in its schema that don't
enforce the actual presence of a vaue in its items. Of interest in this
commit is the `organize` keyword in the core, as well as the `files`
keyword in the copy plugin.

Modify the schema of both the core and the copy plugin to enforce both
the type and the presence of values.

LP: #1587455

Signed-off-by: Kyle Fazzari <kyle@canonical.com>