catkin plugin: produce build-ready staging area.
Currently the Catkin plugin massages some ROS-specific files in order to
get it to build from within the part's install directory. This works
well enough, but produces a staging area that, if used to build more
parts in the future, will attempt to use the old install directory.

Update the Catkin plugin such that, once the build has completed, these
files are massaged to be usable from the staging area.

LP: #1663385

Signed-off-by: Kyle Fazzari <kyle@canonical.com>