plainbox-provider plugin: filter out sitecustomize
When a part that uses the plainbox-provider plugin adds a `stage-packages`
entry for a python3 package, conflicts will occur as this plugin
most likely always comes `after` a python part that actually
adds functionality.

Given that sitecustomize.py will be different and the one added
by `stage-packages` is not really useful in a snap we filter it out.

LP: #1663792

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>