catkin plugin: support building with an underlay.
Currently the Catkin plugin simply pulls down all dependencies of the
workspace being built using rosdep. This doesn't work for the cases of
trying to create a ROS system that utilizes content sharing, as it
forces each snap to bundle their dependencies even when it's undesired.

Add support for an `underlay` property to the Catkin plugin, which
accepts a build-time underlay path for when the consumer snap is being
built, as well as a run-time underlay path for when the consumer snap is
being run (so the workspace can be properly overlaid).

LP: #1664776

Signed-off-by: Kyle Fazzari <kyle@canonical.com>