plainbox-provider plugin: run validate (#1095)
Modify the plainbox-provider plugin to run the "validate" command during the build step to ensure buggy providers don't make it in to checkbox snap packages.

As providers can reference jobs from other providers (i.e. have a dependency on that provider) the environment has a variable PROVIDERPATH populated. Coupled with appropriate use of the "after" keyword this allows "validate" steps to identify that jobs are defined correctly. An integration test to confirm this functionality is included.

LP: #1660692