repo: fixup with python, not sed
After unpacking stage-packages, the repo runs a number of fixup steps.
One of those steps is to rewrite usr/bin/xml2-config and
usr/bin/xslt-config to look within the part instead of on the host.
However, this rewrite is done with sed, which has performance issues as
well as poor side effects when it comes to the plugin environment.

Instead of using sed to fix those files, use python.

LP: #1671675

Signed-off-by: Kyle Fazzari <kyle@canonical.com>