repo: ignore symlinks to libc. (#1174)
libc is not typically staged, as it's always contained in the core snap.
However, Snapcraft leaves symlinks to its libraries from stage-packages
alone and intact. This causes issues when it comes to content sharing
and building against tarred staging areas, as various plugins work
really hard to ensure the snap doesn't include dangling symlinks, and
they end up following the libc symlinks in the tarred staging area and
pulling in libc. The naive solution to this is to simply remove such
symlinks from the stage-packages, and this was indeed the solution
introduced in commit 3052b0f459ded440dad2e2362946ee37e09c8606. However,
didn't work quite as well as desired and led to odd bugs such as ld
linking to static libraries in the part instead of shared libraries on
the host, etc.

Take 2 at a solution: instead of removing symlinks to libc, modify the
copy and dump plugins to ignore (i.e. not follow) symlinks to libc.

LP: #1658774

Signed-off-by: Kyle Fazzari <kyle@canonical.com>