cleanbuild: packaging independent detection.
Don't use the repo to detect if lxd is available as this will prevent use when lxd is provided as a snap or manually installed.

It also makes the logic packaging , agnosting, meaning that a translation layer to figure out what provides the `lxc` command is not necessary.

The work here also migrates the detection out of the lifecycle and into the lxd module.

LP: #1666735
Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>