cleanbuild: don't copy cache into container.
The cache copy logic needs some more design. Currently it copies in the cache regardless of the  target. We need to be smart and only copy the matching hashes. It also needs to weigh in non local remotes and provide user visible feedback in the case where the cache is copied to a non local remote.

LP: #1669730
Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>