pluginhandler: exclude `/snap/` from libraries.
When collecting libraries when working with classic confinement
we currently collect snaps of our own doing if installed. This
small fix avoids that.

Ideally it would have been good to just add the `no-system-libs`
`build-attributes` but the side effects there are a bit too big
as some libraries are still collected when dealing with snaps
that depend on other snaps.

LP: #1670100

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>