lxd: refactor Cleanbuilder into Containerbuild and add Project. (#1230)
Refactored the Cleanbuilder class into a new base class, Containerbuild and introduced a new class, Project, which creates and re-uses a persistent container named after the project (snap). The "snap" as well as the argumentless command now prompts the user for a containerized build  which mounts folders from the host into the container instead of building off a tarball.