cli: provide a whoami command
There are proper fallbacks here as we never
stored the email in our original credentail config location. This provides the proper messaging so this is correct for users wanting to know this.

In some distan future, the account info from the store will also provide this information which is why we loop over the account information .

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>