tests: workaround windowing issue that causes failures during the core snap download
Following the suggestion from William Grant, here we disable the default behaviour for packets outside of the receive window. This should avoid the reset failures we see when downloading the core snap from the Internap CDN during the Travis test.

A better solution from the CDN is coming in July, so this change should be reverted soon.

More information: https://forum.snapcraft.io/t/101-connection-resets/775/19