further attempt to fix cleanbuild and CI scenarios
cleanbuild and CI both run as root and anchor the source tree to /root which is readable only to the root user. because jhbuild runs as a non-root user it failed to access the source because /root was inaccessible. This commit brute forces it's way using chown on the /root folder