project loader: refactor into package
Pave the way for the grammar to be moved into the project loader instead
of pluginhandler. Move `project_loader` into its own package dedicated to
loading the project. Move the `PartsConfig` logic in there as well, but
leave the remote parts logic out. Also, since the schema validator is
only ever used from project loader, move the schema into the
`project_loader` package as well. Finally, create a base class for
project-loading-related errors, and modify the current errors to inherit
from that.

Updates: #1492

Signed-off-by: Kyle Fazzari <kyle@canonical.com>