many: simplify plugin loading (#1507)
Loading a plugin goes through many layers of redirection, which is
confusing and causes components to be integrated more tightly than would
be ideal. Simplify it by no longer loading plugins within PluginHandler,
but rather accepting an already-loaded plugin. That plugin is then
loaded in PartsConfig, where it is then handed to the PluginHandler.

Signed-off-by: Kyle Fazzari <kyle@canonical.com>