project_loader: process stage package grammar
Extract the processing of the stage package grammar from the
pluginhandler and place it into the project_loader. Create an
encapsulating grammar processor dedicated to the grammar in the part
(with the vision of adding grammar support to more part properties in
the future, e.g. `source`). Have the pluginhandler then use this
processor to extract grammar-powered properties, leaving the raw YAML
alone for state tracking.

Updates #1492

Signed-off-by: Kyle Fazzari <kyle@canonical.com>