catkin plugin: don't assume catkin is in underlay
Currently the Catkin plugin pulls its own Catkin off to the side in
order to search for dependencies in the underlay, but then doesn't use
it in isolation, using Catkin components out of the underlay. This
breaks if the underlay doesn't actually include Catkin, and makes no
sense considering that Catkin is already pulled off to the side. Start
using that version completely.

LP: #1696014

Signed-off-by: Kyle Fazzari <kyle@canonical.com>