repo: Use os-release(5) to detect supported Linux distributions
All supported Linux distributions provide an '/etc/os-release' file that
contains the necessary information for determining supported repository
modules to use.

Thus, we can rely on it instead of the deprecated 'linux_distribution'
method from the 'platform' module that will be removed in Python 3.7.

This will also make it much easier to extend to support RPM-based
Linux distributions.

Signed-off-by: Neal Gompa <ngompa13@gmail.com>