dirs: set plugin, schema, and library dir for snap
The default plugin, schema, and library dir are relative to
`sys.prefix`, which works nicely for Snapcraft as a deb or in a venv.
However, Snapcraft's Python plugin installs packages in `$SNAP/` as a
prefix, while Python itself is contained in `$SNAP/usr/`. As a result,
using `sys.prefix` (which is `/usr`) to find these files when Snapcraft
is a snap ends up looking in the wrong place.

Treat running Snapcraft as a snap as a special case, similar to how
running from git is supported, and set the directories appropriately.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>