repo: return a proper value in DummyRepo
The DummyRepo call for get_packages_for_source_type should return
an expected value by the API which is a set.

This happens when snapcraft loads the configuration in cleanbuild to
retrieve the name and some other metadata. It's too early to resolve the
parts and they are null outside of the container.