dotnet plugin: new plugin (#1574)
MVP for a dotnet plugin leveraging existing assets available.
Once the dotnet-sdk and the new json for parsing releases is out,
the plugin would leverage that for built time. Additionally, once the
dotnet-runtime is out, the snaps should allow for a configuration where
the runtime is not provided in the application snap.