snapcraft.yaml: don't re-use build dir (#1601)
Creating the snapcraft snap works fine as long as one is using the snap
or the deb to build. However, for reasons unknown to me, when running
from source, the `make startup` of the `apt` part creates a directory
named `build` that clashes with the `build` scriptlet that also tries to
create a directory named `build`. I'm still not sure why this only
happens when building from source, but an easy fix is to simply create a
different directory in the `build` scriptlet.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>