apt repo: allow insecure repos
In Xenial this was the default behavior, which explains why Snapcraft
works. However, in Zesty this behavior changed, causing breakage when
using external repositories (e.g. ROS). Instead of relying on defaults,
ensure the behavior is consistent across releases.

This is not a long-term fix. We should add a way to ack repository keys.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>