plugins: build-attributes is already in the state
The build state tracks `build-attributes` as a build property, which
means if it changes, the build step is considered dirty. However, a few
plugins have also added `build-attributes` to their build properties,
which is redundant.

Make sure the core Snapcraft properties are handled in the state
classes, and keep the plugins handling their own schema.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>