Removed dependency on VERSION_ID in os-release
_get_system_libs depends on VERSION_ID for the name of the file that is
expected to contain the list of shared libraries to exclude from the build.
However, according to os-release(5), VERSION_ID is just optional, so its
absence should not cause failures.

With this commit, the absence of VERSION_ID is considered as the file
with the list of shared libraries to be excluded does not exist.

Moreover, added dedicated test in snapcraft/tests/test_libraries.py
using Gentoo os-release-info, which does not have VERSION_ID field.

Fixes https://bugs.launchpad.net/snapcraft/+bug/1722650