catkin plugin: support building entire workspace
Currently the Catkin plugin requires one to specify each and every
Catkin package in the workspace that needs to be built and installed
into the snap. This is a hindrance for larger projects, especially as
oftentimes the entire workspace is needed.

Make the `catkin-packages` property optional. If it's not specified,
simply build the entire workspace. Thus the only API change is making
what was previously a required property optional, and assigning meaning
to its absence.

LP: #1721168
Resolve #1713

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>