internal: more gracefully determine host OS (#1636)
Currently snapcraft just reads the `/etc/os-release` file into a dict
and accesses keys everywhere without caring whether or not they exist.
This leads to difficult-to-diagnose errors on distros where e.g. the
`VERSION_CODENAME` isn't set. That particular example includes Trusty,
where we can use the VERSION_ID as a fallback to determine the codename.

Create a class around parsing `/etc/os-release` so we can gracefully
fallback as appropriate. This allows the snapcraft snap to be used on
Trusty, and also makes the errors when running on an unsupported distro
more friendly.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>