autotools: cross-compile using --host instead of env
Currently the autotools plugin uses the CC and CXX environment variables
to force autotools to cross-compile. However, autotools will do this on
its own if one sets the --host parameter correctly, and will actually do
a better job (e.g. use the cross-toolchain's strip as well).

Remove the environment variables, and set the --host to the target
triplet if cross-compiling.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>