catkin plugin: check for pip packages in part only
Currently the Catkin plugin generates a sitecustomize.py if:

  1. pip is setup
  2. pip packages have been pulled

This is good in theory, but pip will be considered setup even if it's
only available in the staging area, and `pip list` will return packages
in the staging area, which means Catkin will pretty much try to generate
a sitecustomize.py if there are any Python parts in the snap, even if it
doesn't use pip itself.

Make sure it checks `pip list --user` rather than just `pip list`. This
will ensure it only looks for pip packages installed by the Catkin part
rather than all parts in the snap.

LP: #1730248

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>