lxd: let lxd choose the architecture
lxd implements exactly the same logic as snapcraft does to choose an
architecture when one is not specified, so remove snapcraft's copy.

It is still necessary to record the architecture of the created
container to detect whether snaps can just be copied across, but this
is now done after the container is started, which will still be correct
when/if support for building on a non-default image is merged.